Split blackbox encoding into separate module
